benchmarks for hex string binary JSON format
